+++ gistit-daemon/src/network.rs
@@ -1,128 +1,61 @@
 //! The network module
 #![allow(clippy::missing_errors_doc)]
 
-use std::collections::HashSet;
-use std::iter::once;
-use std::str::FromStr;
+use std::collections::{HashMap, HashSet};
 use std::string::ToString;
-use std::time::Duration;
+use std::task::Poll;
 
 use either::Either;
 use gistit_ipc::{self, Bridge, Instruction, Server, ServerResponse};
-use log::{debug, info, trace, warn};
-use void::Void;
+use log::{debug, error, info, warn};
 
 use libp2p::core::either::EitherError;
-use libp2p::core::upgrade;
 use libp2p::core::PeerId;
+use libp2p::futures::future::poll_fn;
 use libp2p::futures::StreamExt;
-use libp2p::multiaddr::{multiaddr, Protocol};
-use libp2p::swarm::{ProtocolsHandlerUpgrErr, SwarmEvent};
-use libp2p::{autonat, dns, mplex, noise, tcp, websocket, yamux, Multiaddr, Swarm, Transport};
+use libp2p::multiaddr::multiaddr;
+use libp2p::swarm::{ProtocolsHandlerUpgrErr, SwarmBuilder, SwarmEvent};
+use libp2p::{tokio_development_transport, Multiaddr, Swarm};
 
-use libp2p::identify::{Identify, IdentifyConfig, IdentifyEvent, IdentifyInfo};
-use libp2p::kad::record::store::MemoryStore;
-use libp2p::kad::{
-    self, GetProvidersOk, Kademlia, KademliaConfig, KademliaEvent, QueryId, QueryResult,
-};
-use libp2p::relay::v2::relay;
-use libp2p::request_response::{ProtocolSupport, RequestResponse, RequestResponseConfig};
+use libp2p::identify::{IdentifyEvent, IdentifyInfo};
+use libp2p::kad::{protocol, record::Key, QueryId};
+use libp2p::ping::Failure;
+use libp2p::request_response::RequestId;
 
-use crate::behaviour::{Behaviour, Event, ExchangeCodec, ExchangeProtocol};
+use crate::behaviour::{Behaviour, Event, Request, BOOTADDR};
 use crate::config::Config;
+use crate::event::{handle_kademlia, handle_request_response};
 use crate::Result;
 
 /// The main event loop
 pub struct Node {
-    swarm: Swarm<Behaviour>,
-    bridge: Bridge<Server>,
-    pending_dial: HashSet<PeerId>,
-    pending_start_providing: HashSet<QueryId>,
-    pending_get_providers: HashSet<QueryId>,
-    // pending_request_file: HashMap<RequestId, oneshot::Sender<Result<String>>>,
+    pub swarm: Swarm<Behaviour>,
+    pub bridge: Bridge<Server>,
+
+    pub pending_dial: HashSet<PeerId>,
+
+    /// Pending kademlia queries to get providers
+    pub pending_get_providers: HashSet<QueryId>,
+
+    pub pending_start_providing: HashSet<QueryId>,
+    pub to_provide: HashMap<Key, Vec<u8>>,
+
+    pub pending_request_file: HashSet<RequestId>,
+
+    /// Stack of request file (`key`) events
+    pub to_request: Vec<(Key, HashSet<PeerId>)>,
 }
-
-const BOOTNODES: [&str; 4] = [
-    "QmNnooDu7bfjPFoTZYxMNLWUQJyrVwtbZg5gBMjTezGAJN",
-    "QmQCU2EcMqAqQPR2i9bChDtGNJchTbq5TbXJJ16u19uLTa",
-    "QmbLHAnMoJPWSCR5Zhtx6BHJX9KiKNN6tpvbUcqanj75Nb",
-    "QmcZf59bWwK5XFi76CZX8cbJ4BhTzzA3gU1ZjYZcYW3dwt",
-];
-
-const BOOTADDR: &str = "/dnsaddr/bootstrap.libp2p.io";
 
 impl Node {
     pub async fn new(config: Config) -> Result<Self> {
-        let noise_keys = noise::Keypair::<noise::X25519Spec>::new()
-            .into_authentic(&config.keypair)
-            .expect("Signing libp2p-noise static DH keypair failed.");
+        let behaviour = Behaviour::new(&config)?;
+        let transport = tokio_development_transport(config.keypair)?;
 
-        let transport = {
-            let tcp = tcp::TcpConfig::new().nodelay(true);
-            let dns_tcp = dns::DnsConfig::system(tcp).await?;
-            let ws_dns_tcp = websocket::WsConfig::new(dns_tcp.clone());
-            dns_tcp
-                .or_transport(ws_dns_tcp)
-                .upgrade(upgrade::Version::V1)
-                .authenticate(noise::NoiseConfig::xx(noise_keys).into_authenticated())
-                .multiplex(upgrade::SelectUpgrade::new(
-                    yamux::YamuxConfig::default(),
-                    mplex::MplexConfig::default(),
-                ))
-                .timeout(Duration::from_secs(20))
-                .boxed()
-        };
-
-        let request_response = RequestResponse::new(
-            ExchangeCodec,
-            once((ExchangeProtocol, ProtocolSupport::Full)),
-            RequestResponseConfig::default(),
-        );
-
-        let kademlia = {
-            let mut cfg = KademliaConfig::default();
-            cfg.set_query_timeout(Duration::from_secs(5 * 60));
-            let store = MemoryStore::new(config.peer_id);
-            let mut behaviour = Kademlia::with_config(config.peer_id, store, cfg);
-
-            let bootaddr = Multiaddr::from_str(BOOTADDR)?;
-            for peer in &BOOTNODES {
-                behaviour.add_address(
-                    &PeerId::from_str(peer).expect("peer id to be valid"),
-                    bootaddr.clone(),
-                );
-            }
-
-            behaviour.bootstrap().expect("to bootstrap");
-            behaviour
-        };
-
-        let identify = Identify::new(IdentifyConfig::new(
-            "/ipfs/0.1.0".into(),
-            config.keypair.public(),
-        ));
-
-        let relay = relay::Relay::new(
-            PeerId::from(config.keypair.public()),
-            relay::Config::default(),
-        );
-        let autonat = autonat::Behaviour::new(
-            PeerId::from(config.keypair.public()),
-            autonat::Config::default(),
-        );
-
-        let swarm = Swarm::new(
-            transport,
-            Behaviour {
-                request_response,
-                kademlia,
-                identify,
-                relay,
-                autonat,
-            },
-            config.peer_id,
-        );
-
+        let swarm = SwarmBuilder::new(transport, behaviour, config.peer_id)
+            .executor(Box::new(|fut| {
+                tokio::task::spawn(fut);
+            }))
+            .build();
         let bridge = gistit_ipc::server(&config.runtime_dir)?;
 
         Ok(Self {
@@ -131,6 +64,10 @@ impl Node {
             pending_dial: HashSet::default(),
             pending_start_providing: HashSet::default(),
             pending_get_providers: HashSet::default(),
+            pending_request_file: HashSet::default(),
+
+            to_provide: HashMap::default(),
+            to_request: Vec::default(),
         })
     }
 
@@ -138,13 +75,31 @@ impl Node {
         loop {
             tokio::select! {
                 swarm_event = self.swarm.next() => self.handle_swarm_event(
-                    swarm_event.expect("to recv swarm event")).await?,
+                    swarm_event.expect("stream not to end")).await?,
 
-                bridge_event = async { self.bridge.recv() } => self.handle_bridge_event(bridge_event?).await?
+                bridge_event = self.bridge.recv() => self.handle_bridge_event(bridge_event?).await?,
+
+                request_event = poll_fn(|_| {
+                    self.to_request.pop().map_or(Poll::Pending, Poll::Ready)
+                }) => self.handle_request_event(request_event).await,
             }
         }
     }
 
+    async fn handle_request_event(&mut self, event: (Key, HashSet<PeerId>)) {
+        let (key, providers) = event;
+
+        for p in providers {
+            let request_id = self
+                .swarm
+                .behaviour_mut()
+                .request_response
+                .send_request(&p, Request(key.to_vec()));
+            self.pending_request_file.insert(request_id);
+        }
+    }
+
+    #[allow(clippy::type_complexity)]
     async fn handle_swarm_event(
         &mut self,
         event: SwarmEvent<
@@ -152,17 +107,23 @@ impl Node {
             EitherError<
                 EitherError<
                     EitherError<
-                        EitherError<ProtocolsHandlerUpgrErr<std::io::Error>, std::io::Error>,
-                        std::io::Error,
-                    >,
-                    Either<
-                        ProtocolsHandlerUpgrErr<
-                            EitherError<impl std::error::Error, impl std::error::Error>,
+                        EitherError<
+                            EitherError<ProtocolsHandlerUpgrErr<std::io::Error>, std::io::Error>,
+                            std::io::Error,
                         >,
-                        Void,
+                        Either<
+                            ProtocolsHandlerUpgrErr<
+                                EitherError<
+                                    impl std::error::Error + Send,
+                                    impl std::error::Error + Send,
+                                >,
+                            >,
+                            void::Void,
+                        >,
                     >,
+                    ProtocolsHandlerUpgrErr<std::io::Error>,
                 >,
-                ProtocolsHandlerUpgrErr<std::io::Error>,
+                Failure,
             >,
         >,
     ) -> Result<()> {
@@ -179,7 +140,7 @@ impl Node {
                 debug!("Identify: {:?}", listen_addrs);
                 if protocols
                     .iter()
-                    .any(|p| p.as_bytes() == kad::protocol::DEFAULT_PROTO_NAME)
+                    .any(|p| p.as_bytes() == protocol::DEFAULT_PROTO_NAME)
                 {
                     for addr in listen_addrs {
                         self.swarm
@@ -189,38 +150,26 @@ impl Node {
                     }
                 }
             }
-            //
-            // Kademlia events
-            //
-            SwarmEvent::Behaviour(Event::Kademlia(KademliaEvent::OutboundQueryCompleted {
-                id,
-                result: QueryResult::StartProviding(_),
-                ..
-            })) => {
-                info!("Start providing in kad");
-                self.pending_start_providing.remove(&id);
-            }
-            SwarmEvent::Behaviour(Event::Kademlia(KademliaEvent::OutboundQueryCompleted {
-                id,
-                result: QueryResult::GetProviders(Ok(GetProvidersOk { providers, .. })),
-                ..
-            })) => {
-                info!("Got providers: {:?}", providers);
-                self.pending_get_providers.remove(&id);
-            }
-            SwarmEvent::NewListenAddr { address, .. } => {
-                info!("Daemon: Listening on {:?}", address);
 
+            SwarmEvent::Behaviour(Event::Kademlia(event)) => handle_kademlia(self, event),
+
+            SwarmEvent::Behaviour(Event::RequestResponse(event)) => {
+                handle_request_response(self, event).await?;
+            }
+
+            SwarmEvent::NewListenAddr { address, .. } => {
                 let peer_id = self.swarm.local_peer_id().to_string();
+                info!("Daemon: Listening on {:?}, {:?}", address, peer_id);
 
                 self.bridge.connect_blocking()?;
                 self.bridge
-                    .send(Instruction::Response(ServerResponse::PeerId(peer_id)))?;
+                    .send(Instruction::Response(ServerResponse::PeerId(peer_id)))
+                    .await?;
             }
             SwarmEvent::ConnectionEstablished {
                 peer_id, endpoint, ..
             } => {
-                debug!("Connection established {:?}", peer_id);
+                info!("Connection established {:?}", peer_id);
                 if endpoint.is_dialer() {
                     self.pending_dial.remove(&peer_id);
                 }
@@ -230,13 +179,13 @@ impl Node {
                 error,
                 ..
             } => {
-                debug!("Outgoing connection error: {:?}", error);
+                error!("Outgoing connection error: {:?}", error);
                 if let Some(peer_id) = maybe_peer_id {
                     self.pending_dial.remove(&peer_id);
                 }
             }
             ev => {
-                trace!("other event: {:?}", ev);
+                debug!("other event: {:?}", ev);
             }
         }
         Ok(())
@@ -245,51 +194,59 @@ impl Node {
     async fn handle_bridge_event(&mut self, instruction: Instruction) -> Result<()> {
         match instruction {
             Instruction::Listen { host, port } => {
-                info!("Instruction: Listen");
-                let addr = multiaddr!(Ip4(host), Tcp(port));
-                self.swarm.listen_on(addr)?;
+                warn!("Instruction: Listen");
+                let address = multiaddr!(Ip4(host), Tcp(port));
+                self.swarm.listen_on(address)?;
             }
-            Instruction::Dial { peer_id } => {
-                info!("Instruction: Dial");
 
-                // let addr: Multiaddr = BOOTADDR.parse().unwrap();
-                let addr: Multiaddr = "/ip4/192.168.1.77/tcp/4001".parse().unwrap();
+            Instruction::Dial { peer_id } => {
+                warn!("Instruction: Dial");
+
+                let base_addr: Multiaddr = BOOTADDR.parse().unwrap();
                 let peer: PeerId = peer_id.parse().unwrap();
 
                 if self.pending_dial.contains(&peer) {
-                    debug!("Already dialing peer: {}", peer_id);
+                    error!("Already dialing peer: {}", peer_id);
                     return Ok(());
                 }
 
                 self.swarm
                     .behaviour_mut()
                     .kademlia
-                    .add_address(&peer, addr.clone());
+                    .add_address(&peer, base_addr);
 
-                self.swarm.dial(addr.with(Protocol::P2p(peer.into())))?;
+                // self.swarm
+                //     .dial()?;
                 self.pending_dial.insert(peer);
             }
-            Instruction::Provide { hash, .. } => {
-                info!("Instruction: Provide file {}", hash);
 
-                self.swarm
-                    .behaviour_mut()
-                    .kademlia
-                    .start_providing(hash.into_bytes().into())
-                    .expect("to start providing");
-            }
-            Instruction::Get { hash } => {
-                info!("Instruction: Get providers for {}", hash);
+            Instruction::Provide { hash, data } => {
+                warn!("Instruction: Provide gistit {}", hash);
+                let key = Key::new(&hash);
 
                 let query_id = self
                     .swarm
                     .behaviour_mut()
                     .kademlia
-                    .get_providers(hash.into_bytes().into());
+                    .start_providing(key.clone())
+                    .expect("to start providing");
+
+                self.pending_start_providing.insert(query_id);
+                self.to_provide.insert(key, data);
+            }
+
+            Instruction::Get { hash } => {
+                warn!("Instruction: Get providers for {}", hash);
+                let query_id = self
+                    .swarm
+                    .behaviour_mut()
+                    .kademlia
+                    .get_providers(Key::new(&hash));
                 self.pending_get_providers.insert(query_id);
             }
+
             Instruction::Status => {
-                info!("Instruction: Status");
+                warn!("Instruction: Status");
 
                 let listeners: Vec<String> =
                     self.swarm.listeners().map(ToString::to_string).collect();
@@ -300,12 +257,15 @@ impl Node {
                     .send(Instruction::Response(ServerResponse::Status(format!(
                         "listeners: {:?}, network: {:?}",
                         listeners, network_info
-                    ))))?;
+                    ))))
+                    .await?;
             }
+
             Instruction::Shutdown => {
                 warn!("Exiting...");
                 std::process::exit(0);
             }
+
             _ => (),
         }
         Ok(())